requirements.txt
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
beautifulsoup4==4.12.2
billboard.py==7.0.0
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dacite==1.8.1
Deprecated==1.2.14
exceptiongroup==1.2.0
fastapi==0.103.2
h11==0.14.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
jaconv==0.3.4
lxml==5.1.0
lyrics-extractor==3.0.1
lyricsgenius==3.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
mutagen==1.47.0
pillow==10.2.0
platformdirs==3.11.0
pycryptodomex==3.20.0
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
pykakasi==2.2.1
pyTelegramBotAPI==4.14.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.3
pytube==15.0.0
rapidfuzz==3.6.1
redis==5.0.1
requests==2.31.0
rich==13.7.0
six==1.16.0
sniffio==1.3.0
soundcloud-v2==1.3.1
soupsieve==2.5
spotdl==4.2.4
spotipy==2.23.0
starlette==0.27.0
syncedlyrics==0.7.0
telebot==0.0.5
text-unidecode==1.3
tqdm==4.66.1
typing_extensions==4.9.0
Unidecode==1.3.8
urllib3==2.2.0
uvicorn==0.23.2
websockets==12.0
wrapt==1.16.0
yt-dlp==2023.12.30
ytmusicapi==1.5.1
zipp==3.17.0
